refactor(weather): Flatten weather service directory structure
- Moved service files out of nested services/ into a single weather/ directory
- Adjusted import paths accordingly
- Applies to: IWeatherApiClient, StubWeatherDataSource, WeatherApiClient, and provider
